backend/cli_service/requirements.txt
typer==0.12.3
httpx==0.27.0
rich==13.7.1
pyyaml==6.0.1
pydantic-settings==2.2.1
python-jose[cryptography]==1.3.0
